Shows alert screen when a fall is detected and sends alert if needed
