+++ mobile/src/main/java/gameofphones/gatech/stumble/AlertWhoActivity.java
@@ -1,27 +1,19 @@
 package gameofphones.gatech.stumble;
 
-import android.Manifest;
 import android.app.Activity;
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.content.pm.PackageManager;
-import android.os.Build;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.telephony.SmsManager;
 import android.util.Log;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.Toast;
-
-import static android.Manifest.permission.SEND_SMS;
 
 public class AlertWhoActivity extends AppCompatActivity {
 
@@ -34,9 +26,6 @@ public class AlertWhoActivity extends AppCompatActivity {
     private BroadcastReceiver sentStatusReceiver, deliveredStatusReceiver;
 
     private String phoneNumber;
-
-    private static final int REQUEST_SMS = 0;
-
 
     ListView listView;
 
@@ -56,34 +45,10 @@ public class AlertWhoActivity extends AppCompatActivity {
             isTest = true;
         }
 
-
-
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position,
                                     long id) {
-                if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
-                    int hasSMSPermission = checkSelfPermission(Manifest.permission.SEND_SMS);
-                    if (hasSMSPermission != PackageManager.PERMISSION_GRANTED) {
-                        if (!shouldShowRequestPermissionRationale(Manifest.permission.SEND_SMS)) {
-                            showMessageOKCancel("You need to allow access to Send SMS",
-                                    new DialogInterface.OnClickListener() {
-                                        @Override
-                                        public void onClick(DialogInterface dialog, int which) {
-                                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                                                requestPermissions(new String[]{Manifest.permission.SEND_SMS},
-                                                        REQUEST_SMS);
-                                            }
-                                        }
-                                    });
-                            return;
-                        }
-                        requestPermissions(new String[]{Manifest.permission.SEND_SMS},
-                                REQUEST_SMS);
-                        return;
-                    }
-                    sendFall();
-                }
                 sendFall();
                 Intent intent = new Intent(AlertWhoActivity.this, AlertSentActivity.class);
                 String message = templateNamesArr[position];
@@ -92,7 +57,6 @@ public class AlertWhoActivity extends AppCompatActivity {
             }
         });
     }
-
 
     public void sendFall() {
         //will allow text message to be sent to a contact
@@ -115,15 +79,6 @@ public class AlertWhoActivity extends AppCompatActivity {
             PendingIntent deliveredIntent = PendingIntent.getBroadcast(this, 0, new Intent("SMS_DELIVERED"), 0);
             sms.sendTextMessage(contactNumber, null, alertMessage, sentIntent, deliveredIntent);
         }
-    }
-
-    private void showMessageOKCancel(String message, DialogInterface.OnClickListener okListener) {
-        new android.support.v7.app.AlertDialog.Builder(AlertWhoActivity.this)
-                .setMessage(message)
-                .setPositiveButton("OK", okListener)
-                .setNegativeButton("Cancel", null)
-                .create()
-                .show();
     }
 
     public void onResume() {
@@ -177,44 +132,5 @@ public class AlertWhoActivity extends AppCompatActivity {
         super.onPause();
         unregisterReceiver(sentStatusReceiver);
         unregisterReceiver(deliveredStatusReceiver);
-    }
-
-    private boolean checkPermission() {
-        return ( ContextCompat.checkSelfPermission(getApplicationContext(), SEND_SMS ) == PackageManager.PERMISSION_GRANTED);
-    }
-
-    private void requestPermission() {
-        ActivityCompat.requestPermissions(this, new String[]{SEND_SMS}, REQUEST_SMS);
-    }
-
-    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
-        switch (requestCode) {
-            case REQUEST_SMS:
-                if (grantResults.length > 0 &&  grantResults[0] == PackageManager.PERMISSION_GRANTED){
-                    Toast.makeText(getApplicationContext(), "Permission Granted, Now you can access sms", Toast.LENGTH_SHORT).show();
-                    sendFall();
-
-                }else {
-                    Toast.makeText(getApplicationContext(), "Permission Denied, You cannot access and sms", Toast.LENGTH_SHORT).show();
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                        if (shouldShowRequestPermissionRationale(SEND_SMS)) {
-                            showMessageOKCancel("You need to allow access to both the permissions",
-                                    new DialogInterface.OnClickListener() {
-                                        @Override
-                                        public void onClick(DialogInterface dialog, int which) {
-                                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                                                requestPermissions(new String[]{SEND_SMS},
-                                                        REQUEST_SMS);
-                                            }
-                                        }
-                                    });
-                            return;
-                        }
-                    }
-                }
-                break;
-            default:
-                super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        }
     }
 }
